book-recommender/requirements.txt
kagglehub
pandas
matplotlib
seaborn
python-dotenv
langchain-community 
langchain-openai 
langchain-chroma
transformers
gradio
notebook 
ipywidgets


